Fix dangling link by pointing to OpenShift 3.11 console walkthrough
The target doesn't exist anymore on the current "latest" tree
(corresponding to OpenShift 4.1).  As long as most APPUiO
installations in the real world are based on OpenShift 3.11 or older,
it is better to link to that "old" version of the page than to a
non-existing page or even a variant of this documentation that may not
exist to what users actually see.